fix(etsy): return-policy label + settings as its own route; consume plugin @latest
- Etsy return policies have no `name` (only accepts_returns / accepts_exchanges /
  return_deadline), so the Sync-defaults select rendered blank → showed "Not set"
  even though the API returned a policy. getShopReturnPolicies now derives a
  readable name (describeReturnPolicy → e.g. "Returns in 30 days + exchanges");
  the select also falls back to the id.
- Extract the Sync-settings drawer out of the Etsy list page into its own route
  (settings/etsy/settings) — a routed drawer that navigates back to the list on
  close. Trims the list page to connection + recent-syncs.
- Bump plugin to 0.1.3 and switch apps/backend to consume "@latest" instead of a
  pinned/incremented version (lockfile specifier reconciled, resolves to 0.1.2
  until 0.1.3 is published).

Plugin tsc: 0 errors.
