+++ apps/backend/package.json
@@ -68,7 +68,7 @@
     "@freesewing/titan": "4.10.0",
     "@hookform/error-message": "^2.0.1",
     "@hookform/resolvers": "^5.0.1",
-    "@jytextiles/medusa-plugin-etsy-sync": "0.1.2",
+    "@jytextiles/medusa-plugin-etsy-sync": "latest",
     "@mastra/core": "latest",
     "@mastra/evals": "latest",
     "@mastra/loggers": "latest",
